requirements.txt
wheel
celery==5.2.1
Django==3.2
django_celery_beat==2.2.1
django-default-imagefield==0.5.5
django-import-export==3.2.0
django-notifications-hq==1.7.0
django-phonenumber-field==7.0.0
django-sass-processor==1.2.2
django-session-timeout==0.1.0
django-widget-tweaks==1.4.12
email-validator==1.3.0
fpdf2==2.7.7
hiredis==2.0.0
libsass==0.21.0
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2==2.9.5
pyjwt==2.6.0
python-decouple==3.6
redis==4.0.1
requests==2.28.1
uwsgi==2.0.21